Clean up around index testing
- Tests that verify that index are being planned should live in `LeafPlanningIntegrationTests`
- Tests that test the actual index implementation should live in `NodeIndexSeekAcceptanceTest`,
    `NodeIndexScanAcceptanceTest`, or `NodeIndexSeekByRangeAcceptanceTest`
