+++ community/cypher/acceptance/src/test/scala/org/neo4j/internal/cypher/acceptance/LikeAcceptanceTest.scala
@@ -43,232 +43,6 @@ class LikeAcceptanceTest extends ExecutionEngineFunSuite with QueryStatisticsTes
   }
 
   // *** TESTS OF PREFIX SEARCH
-
-  test("should be case sensitive") {
-    val london = createLabeledNode(Map("name" -> "London"), "Location")
-    createLabeledNode(Map("name" -> "london"), "Location")
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Location")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("name" -> i.toString), "Location")
-      }
-    }
-
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Lon%' RETURN l"
-
-    val result = executeWithAllPlanners(query)
-
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("l" -> london)))
-  }
-
-  test("should perform prefix search in an update query") {
-    createLabeledNode(Map("name" -> "London"), "Location")
-    createLabeledNode(Map("name" -> "london"), "Location")
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Lon%' CREATE (L:Location {name: toUpper(l.name)}) RETURN L.name AS NAME"
-
-    val result = executeWithRulePlanner(query)
-
-    result.executionPlanDescription().toString should include("SchemaIndex")
-    result.executionPlanDescription().toString should include("PrefixSeekRange")
-    result.toList should equal(List(Map("NAME" -> "LONDON")))
-  }
-
-  test("should perform prefix search for _ in an update query") {
-    createLabeledNode(Map("name" -> "Loony"), "Location")
-    createLabeledNode(Map("name" -> "loony"), "Location")
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Loon_' CREATE (L:Location {name: toUpper(l.name)}) RETURN L.name AS NAME"
-
-    val result = executeWithRulePlanner(query)
-
-    result.executionPlanDescription().toString should include("SchemaIndex")
-    result.executionPlanDescription().toString should include("PrefixSeekRange")
-    result.toList should equal(List(Map("NAME" -> "LOONY")))
-  }
-
-  test("should perform prefix search for _ in an update query with complex prefix") {
-    createLabeledNode(Map("name" -> "Loonyboom"), "Location")
-    createLabeledNode(Map("name" -> "loonyboom"), "Location")
-    createLabeledNode(Map("name" -> "boom"), "Location")
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Loon_boom' CREATE (L:Location {name: toUpper(l.name)}) RETURN L.name AS NAME"
-
-    val result = executeWithRulePlanner(query)
-
-    result.executionPlanDescription().toString should include("SchemaIndex")
-    result.executionPlanDescription().toString should include("PrefixSeekRange")
-    result.toList should equal(List(Map("NAME" -> "LOONYBOOM")))
-  }
-
-  test("should perform complex prefix search in an update query)") {
-    val london = createLabeledNode(Map("name" -> "London"), "Location")
-    createLabeledNode(Map("name" -> "Londinium"), "Location")
-    createLabeledNode(Map("name" -> "london"), "Location")
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Lon%don' CREATE (L:Location {name: toUpper(l.name)}) RETURN L.name AS NAME"
-
-    val result = executeWithRulePlanner(query)
-
-    result.executionPlanDescription().toString should include("SchemaIndex")
-    result.executionPlanDescription().toString should include("PrefixSeekRange")
-    result.executionPlanDescription().toString should include("Filter")
-    result.toList should equal(List(Map("NAME" -> "LONDON")))
-  }
-
-  test("should only match on the actual prefix") {
-    val london = createLabeledNode(Map("name" -> "London"), "Location")
-    graph.inTx {
-      createLabeledNode(Map("name" -> "Johannesburg"), "Location")
-      createLabeledNode(Map("name" -> "Paris"), "Location")
-      createLabeledNode(Map("name" -> "Malmo"), "Location")
-      createLabeledNode(Map("name" -> "Loondon"), "Location")
-      createLabeledNode(Map("name" -> "Lolndon"), "Location")
-
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Location")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("name" -> i.toString), "Location")
-      }
-    }
-    graph.createIndex("Location", "name")
-
-    val query = "MATCH (l:Location) WHERE l.name LIKE 'Lon%' RETURN l"
-
-    val result = executeWithAllPlanners(query)
-
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("l" -> london)))
-  }
-
-  test("should plan the leaf with the longest prefix if multiple LIKE patterns") {
-
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Address")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("prop" -> i.toString), "Address")
-      }
-    }
-
-    val a1 = createLabeledNode(Map("prop" -> "www123"), "Address")
-    val a2 = createLabeledNode(Map("prop" -> "www"), "Address")
-    createLabeledNode(Map("prop" -> "ww"), "Address")
-
-    graph.createIndex("Address", "prop")
-
-    val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'w%' AND a.prop LIKE 'www%' RETURN a")
-
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.executionPlanDescription().toString should include("prop LIKE www%")
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
-  }
-
-  test("should plan an IndexRangeSeek for a % string prefix search when index exists") {
-
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Address")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("prop" -> i.toString), "Address")
-      }
-    }
-
-    val a1 = createLabeledNode(Map("prop" -> "www123"), "Address")
-    val a2 = createLabeledNode(Map("prop" -> "www"), "Address")
-    val a3 = createLabeledNode(Map("prop" -> "ww"), "Address")
-
-    graph.createIndex("Address", "prop")
-
-    val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'www%' RETURN a")
-
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
-  }
-
-  test("should plan an IndexRangeSeek for a _ string prefix search when index exists") {
-
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Address")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("prop" -> i.toString), "Address")
-      }
-    }
-
-    val a1 = createLabeledNode(Map("prop" -> "www123"), "Address")
-    val a2 = createLabeledNode(Map("prop" -> "www"), "Address")
-    val a3 = createLabeledNode(Map("prop" -> "ww"), "Address")
-
-    graph.createIndex("Address", "prop")
-
-    val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'ww_' RETURN a")
-
-    result.executionPlanDescription().toString should include("Filter")
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("a" -> a2)))
-  }
-
-  test("should plan an IndexRangeSeek for a string search that starts with a prefix when index exists") {
-
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Address")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("prop" -> i.toString), "Address")
-      }
-    }
-    val a1 = createLabeledNode(Map("prop" -> "www123"), "Address")
-    val a2 = createLabeledNode(Map("prop" -> "www"), "Address")
-    val a3 = createLabeledNode(Map("prop" -> "ww"), "Address")
-
-    graph.createIndex("Address", "prop")
-
-    val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'ww%w%' RETURN a")
-
-    result.executionPlanDescription().toString should include(IndexSeekByRange.name)
-    result.executionPlanDescription().toString should include("Filter")
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
-  }
-
-
-  test("should plan a UniqueIndexSeek when constraint exists") {
-
-    graph.inTx {
-      (1 to 100).foreach { _ =>
-        createLabeledNode("Address")
-      }
-      (1 to 300).map { i =>
-        createLabeledNode(Map("prop" -> i.toString), "Address")
-      }
-    }
-
-    val a1 = createLabeledNode(Map("prop" -> "www123"), "Address")
-    val a2 = createLabeledNode(Map("prop" -> "www"), "Address")
-    val a3 = createLabeledNode(Map("prop" -> "ww"), "Address")
-
-    graph.createConstraint("Address", "prop")
-
-    val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'www%' RETURN a")
-
-    result.executionPlanDescription().toString should include(UniqueIndexSeekByRange.name)
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
-  }
-
   test("should not plan an IndexSeek when index doesn't exist") {
 
     (1 to 10).foreach { _ =>
@@ -280,8 +54,7 @@ class LikeAcceptanceTest extends ExecutionEngineFunSuite with QueryStatisticsTes
 
     val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'www%' RETURN a")
 
-    result.executionPlanDescription().toString should not include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
+    result should (not(use(IndexSeekByRange.name)) and evaluateTo(List(Map("a" -> a1), Map("a" -> a2))))
   }
 
   test("should not plan an IndexSeek when the LIKE pattern does not have a valid prefix") {
@@ -297,8 +70,7 @@ class LikeAcceptanceTest extends ExecutionEngineFunSuite with QueryStatisticsTes
 
     val result = executeWithAllPlanners("MATCH (a:Address) WHERE a.prop LIKE 'ww_%' RETURN a")
 
-    result.executionPlanDescription().toString should not include(IndexSeekByRange.name)
-    result.toList should equal(List(Map("a" -> a1), Map("a" -> a2)))
+    result should (not(use(IndexSeekByRange.name)) and evaluateTo(List(Map("a" -> a1), Map("a" -> a2))))
   }
 
   test("finds exact matches") {
