+++ community/cypher/cypher-compiler-2.3/src/test/scala/org/neo4j/cypher/internal/compiler/v2_3/planner/logical/LeafPlanningIntegrationTest.scala
@@ -19,7 +19,7 @@
  */
 package org.neo4j.cypher.internal.compiler.v2_3.planner.logical
 
-import org.neo4j.cypher.internal.compiler.v2_3.{RangeBetween, _}
+import org.neo4j.cypher.internal.compiler.v2_3._
 import org.neo4j.cypher.internal.compiler.v2_3.ast.{InequalitySeekRangeWrapper, PrefixSeekRangeWrapper}
 import org.neo4j.cypher.internal.compiler.v2_3.commands.{ManyQueryExpression, RangeQueryExpression, SingleQueryExpression}
 import org.neo4j.cypher.internal.compiler.v2_3.pipes.LazyLabel
@@ -30,7 +30,7 @@ import org.neo4j.cypher.internal.compiler.v2_3.planner.logical.plans._
 import org.neo4j.cypher.internal.frontend.v2_3.ast._
 import org.neo4j.cypher.internal.frontend.v2_3.helpers.NonEmptyList
 import org.neo4j.cypher.internal.frontend.v2_3.test_helpers.CypherFunSuite
-import org.neo4j.cypher.internal.frontend.v2_3._
+import org.neo4j.cypher.internal.frontend.v2_3.{ExclusiveBound, InclusiveBound, LabelId, PropertyKeyId}
 
 class LeafPlanningIntegrationTest extends CypherFunSuite with LogicalPlanningTestSupport2 {
 
@@ -553,5 +553,25 @@ class LeafPlanningIntegrationTest extends CypherFunSuite with LogicalPlanningTes
         Map(), Map("count(*)" -> CountStar()_)
       )(solved)
     )
+  }
+
+  test("should use index on label and property") {
+    val plan = new given {
+      indexOn("Crew", "name")
+    } planFor "MATCH (n:Crew) WHERE n.name = 'Neo' RETURN n"
+
+    plan shouldBe using[NodeIndexSeek]
+  }
+
+  test("should use index when there are multiple labels on the node") {
+    val plan = new given {
+      indexOn("Crew", "name")
+      cost = {
+        case (_: NodeByIdSeek, _) => 1.0
+        case _ => 100.0
+      }
+    } planFor "MATCH (n:Matrix:Crew) WHERE n.name = 'Neo' RETURN n"
+
+    plan shouldBe using[NodeIndexSeek]
   }
 }
